% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{boxfivenum9}
\alias{boxfivenum9}
\title{计算带有加权的五数概括（九数概括）}
\usage{
boxfivenum9(data = 1:20)
}
\arguments{
\item{data}{向量类型的数据，默认为1到20的整数向量。用于计算五数概括的输入数据。}
}
\value{
返回一个包含九个值的向量，分别是：
1. 最小值
2. 下四分位数
3. 中位数
4. 加数1
5. 加数2
6. 上四分位数
7. 加数3
8. 加数4
9. 最大值
}
\description{
此函数基于R语言的五数概括（min, Q1, median, Q3, max）计算一个扩展的九数概括。
该函数首先通过1.5倍四分位距计算上下限，并根据这些限制调整五数概括中的最大值和最小值。
然后根据上四分位数和下四分位数的差距，计算加权的中位数加值，生成九数概括。
}
\examples{
boxfivenum9(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
boxfivenum9(1:100)
}
