% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{calculate_growth_rate_for_list}
\alias{calculate_growth_rate_for_list}
\title{计算增长率}
\usage{
calculate_growth_rate_for_list(xts_list)
}
\arguments{
\item{xts_list}{一个列表，每个元素是一个xts对象，包含时间序列数据。}
}
\value{
返回一个数据框，其中包括每个xts对象的名称、首次价格、末次价格和增长率。
}
\description{
该函数接收一个包含多个xts对象的列表，计算每个xts对象的增长率。
对于每个xts对象，函数会提取第一个和最后一个价格，并计算增长率。
}
\examples{
# 假设有一个包含xts对象的列表xts_list
results <- calculate_growth_rate_for_list(xts_list)
print(results)
}
