% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{check_in_out}
\alias{check_in_out}
\title{检查股票在指定日期范围内的买入和卖出情况}
\usage{
check_in_out(
  stock_data_ss50 = mydata,
  before = 1,
  current = 0,
  stock_type = "strongsell"
)
}
\arguments{
\item{stock_data_ss50}{数据框或xts对象，包含股票的历史数据，默认使用 `mydata` 数据集。}

\item{before}{前一个时间段的天数，默认为1，表示在当前时间之前的1天。}

\item{current}{当前时间段的天数，默认为0，表示当前时间。}

\item{stock_type}{股票的买入或卖出类型，默认为 "strongsell"，表示强烈卖出类型。}
}
\value{
返回一个列表 `mylist`，其中包含以下三个元素：
\item{common}{在两个时间段内都有出现的股票代码。}
\item{in_}{在当前时间段内出现，但在前一个时间段内未出现的股票代码。}
\item{out_}{在前一个时间段内出现，但在当前时间段内未出现的股票代码。}
}
\description{
该函数根据输入的股票数据和指定的日期范围，
检查在前一段时间（before）和当前时间（current）内，
指定类型的股票（如“strongsell”）是否在两个时间段内都有出现，
或者在某一时间段内出现。
}
\examples{
\dontrun{
# 假设已经加载了 stock_data_ss50 数据
result <- check_in_out(stock_data_ss50 = mydata, before = 5, current = 0, stock_type = "strongsell")
print(result)
}
}
