% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{deleFromEnd}
\alias{deleFromEnd}
\title{从股票数据中删除最后 n 天数据}
\usage{
deleFromEnd(sz50, deleN = 0)
}
\arguments{
\item{deleN}{一个整数，表示要删除的天数。如果 deleN 为 0，则返回原始数据。}

\item{stock_data_ss50}{一个列表，每个元素是一个时间序列数据（例如 xts 对象），表示股票的每日数据。}
}
\value{
返回一个列表，包含每个股票去除最后 n 天的数据。
}
\description{
该函数从传入的股票数据中删除最后 n 天的数据。
}
\examples{
# 假设 stock_data_ss50 是一个包含股票数据的列表
# 删除每个股票数据的最后 5 天数据
result <- deleFromEnd(stock_data_ss50, deleN = 5)

}
