% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock.R
\name{drawStockLine}
\alias{drawStockLine}
\title{画曲线图}
\usage{
drawStockLine(ssm = stock_data_ss50[[2]])
}
\arguments{
\item{ssm}{股票数据，默认为 `stock_data_ss50[[2]]`。函数会将其转换为数据框，并从行名中提取日期信息。}
}
\value{
返回一个使用 `ggplot2` 包创建的曲线图对象，图中包含股票调整价格的折线图以及两种平滑曲线（GLM 和 GAM）。
}
\description{
该函数用于绘制股票数据的曲线图，并添加不同方法的平滑曲线。
}
\examples{
\dontrun{
# 假设 stock_data_ss50 已经定义
plot <- drawStockLine(stock_data_ss50[[2]])
print(plot)
}

}
