% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock.R
\name{getListStocks}
\alias{getListStocks}
\title{获取多个股票的历史数据}
\usage{
getListStocks(
  stocklist = c("600104.SS", "600028.SS"),
  start = "2020-01-01",
  end = Sys.Date()
)
}
\arguments{
\item{stocklist}{一个字符向量，包含需要获取历史数据的股票代码。默认为 \code{c("600104.SS", "600028.SS")}。}

\item{start}{一个日期字符串，表示获取历史数据的开始日期。默认为 \code{"2020-01-01"}。}

\item{end}{一个日期字符串，表示获取历史数据的结束日期。默认为当前日期（\code{Sys.Date()}）。}
}
\value{
返回一个列表，其中每个元素对应一个股票的历史数据，列表的名字是股票的代码，每个元素是一个数据框或时间序列对象。
        每个数据框包含该股票的历史数据，并添加了一个名为 \code{name} 的列，表示股票的名称。
}
\description{
该函数根据提供的股票代码列表（stocklist）获取多个股票的历史数据，并返回一个包含股票数据的列表。每个列表项包含一个股票的时间序列数据和该股票的名称。
}
\examples{
\dontrun{
stock_data <- getListStocks(c("600104.SS", "600028.SS"))
}

}
