% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock.R
\name{getStocks}
\alias{getStocks}
\title{获取股票数据}
\usage{
getStocks(stock_code, start_date = "2020-01-01", end_date = Sys.Date())
}
\arguments{
\item{stock_code}{一个字符型参数，指定股票代码，例如 "AAPL" 或 "000001.SS"。}

\item{start_date}{一个字符型参数，指定开始日期，格式为 "YYYY-MM-DD"（默认为 "2020-01-01"）。}

\item{end_date}{一个字符型参数，指定结束日期，格式为 "YYYY-MM-DD"（默认为当前日期）。}
}
\value{
返回一个 xts 对象，包含指定股票的历史数据，包括开盘价、最高价、最低价、收盘价、成交量等。
}
\description{
这个函数从 Yahoo Finance 获取指定股票代码的历史数据。
用户可以指定股票代码、开始日期和结束日期来获取相应的股票历史数据。
}
\examples{
\dontrun{
stock_data <- getStocks("AAPL", "2020-01-01", "2021-01-01")
}
}
