% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{judge_buy_sell}
\alias{judge_buy_sell}
\title{判断买入卖出信号}
\usage{
judge_buy_sell(vd = c(6, 6, 7, 8, 9, 10, 9, 8, 7, 7, 7, 6))
}
\arguments{
\item{vd}{一个数值向量，表示价格数据序列。}
}
\value{
返回一个长度与输入向量相同的向量，第一个值为 NA，后续值为对应的信号值（-1 或 0）。
}
\description{
该函数用于根据输入的价格数据判断买入和卖出信号。具体规则为：
- 如果当前值小于等于6，下一值大于等于7，判断为买入信号（-1）。
- 如果当前值大于等于9，下一值小于等于8，判断为卖出信号（0）。
- 如果当前值大于等于10，下一值小于等于9，判断为卖出信号（0）。
}
\examples{
judge_buy_sell(c(6,6,7,8,9,10,9,8,7,7,7,6))
}
