% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock.R
\name{plot_hw}
\alias{plot_hw}
\title{绘制 Holt-Winters 预测图}
\usage{
plot_hw(
  data = stock_data_ss50[1:2],
  prediction_period = 30,
  filename = "hw_plots.html"
)
}
\arguments{
\item{data}{包含多个股票数据的列表。每个元素是一个 xts 对象，表示股票的历史数据。}

\item{prediction_period}{预测的周期，默认为 30 天。}

\item{filename}{生成的 HTML 文件名，默认为 "hw_plots.html"。}
}
\value{
返回保存预测图的 HTML 文件。
}
\description{
该函数接收一个包含多个股票数据的列表，为每个股票数据使用 Holt-Winters 模型进行预测，并绘制交互式图表。
}
\examples{
# 假设有一个包含 xts 对象的股票数据列表 stock_data
plot_hw(data = stock_data, filename = "hw_stock_plots.html")
}
