% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock.R
\name{plot_prophet}
\alias{plot_prophet}
\title{绘制并保存多个股票数据的 Prophet 预测交互式图表到 HTML 文件}
\usage{
plot_prophet(data = stock_data_ss50, filename = "fifty_plots.html")
}
\arguments{
\item{data}{一个列表，列表中的每个元素代表一个股票的时间序列数据。
列表元素的名称将作为图表的标题。默认为 `stock_data_ss50`。}

\item{filename}{字符型，指定要保存的 HTML 文件的名称。默认为 `"fifty_plots.html"`。}
}
\value{
无返回值，函数的主要作用是将生成的 HTML 文件保存到指定路径。
}
\description{
该函数接收包含多个股票数据的列表，为每个股票数据拟合 Prophet 模型进行预测，
生成交互式图表，并将所有图表整合到一个 HTML 文件中，同时为每个图表添加对应的股票名称作为标题。
}
