% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock.R
\name{plot_stock}
\alias{plot_stock}
\title{绘制股票分析图（包含残差分析和蜡烛图）}
\usage{
plot_stock(stockname = "600048.SS")
}
\arguments{
\item{stockname}{字符串类型，表示股票代码，默认为 \code{"600048.SS"}。}
}
\value{
无返回值。函数生成一个 PDF 文件，包含蜡烛图、残差的箱线图和直方图。
}
\description{
该函数获取指定股票的历史数据，进行线性回归残差分析，并生成蜡烛图，保存为 PDF 文件。
}
\details{
函数获取指定股票的日频历史数据，并对其进行分析：
\itemize{
  \item 蜡烛图：显示股票的开盘、最高、最低和收盘价的变化趋势。
  \item 箱线图：展示线性回归残差的分布特征。
  \item 直方图：展示残差的分布情况。
}
}
\examples{
\dontrun{
plot_stock("600048.SS")
}

}
