% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{satstock}
\alias{satstock}
\title{更新和保存股票数据}
\usage{
satstock(my = statable)
}
\arguments{
\item{my}{数据框，默认值为 `statable`，表示需要更新的数据。此数据框应包含股票数据，并有一个 `Date` 列。}
}
\value{
返回更新后的 `saved_statable` 数据框，包含去重和排序后的股票数据。
}
\description{
该函数将新的股票数据添加到已有的数据框 `saved_statable` 中，并去除重复的日期项，按日期排序后保存。
}
\examples{
updated_data <- satstock(new_data)
}
