% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{sta_type_number}
\alias{sta_type_number}
\title{统计不同类型的股票数据数量}
\usage{
sta_type_number(sz50 = stock_data_ss50)
}
\arguments{
\item{sz50}{一个包含股票数据的列表，默认为 `stock_data_ss50`。列表中的每个元素应该是一个数据框或可处理的数据结构，
能够被 `getsellpoint` 函数处理。}
}
\value{
一个数据框，包含日期和不同类型股票数据的统计数量。列名包括 `Date`、`strongbuy`、`buy`、`holdbuy`、
        `holdselldown`、`holdsell`、`holdsellup`、`selldown`、`sell`、`sellup` 和 `strongsell`。
}
\description{
该函数用于对输入的股票数据列表进行处理，计算每种类型的股票数据数量，并将结果整理成一个数据框返回。
它会对输入列表中的每个元素应用 `getsellpoint` 函数，提取最后一行数据，然后统计不同类型的出现次数。
}
\examples{
\dontrun{
result <- sta_type_number(sz50 = stock_data_ss50)
print(result)
}
}
