% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_stock.R
\name{stock_prediction_prophet}
\alias{stock_prediction_prophet}
\title{股票价格预测函数}
\usage{
stock_prediction_prophet(
  stock_data = stock_data_ss50[["600048.SS"]],
  prediction_period = 30
)
}
\arguments{
\item{stock_data}{数据框，包含股票历史数据。默认值为 `stock_data_ss50[["600048.SS"]]`。
数据框中需包含以下列：
  - `Date`：日期列，表示股票的交易日期。
  - `Adjusted`：调整后收盘价。}

\item{prediction_period}{整数，表示预测的天数。默认值为 30。}
}
\value{
返回一个 xts 对象，包含预测的未来日期及相应的预测价格。
}
\description{
使用 Prophet 模型对股票数据进行预测。此函数读取股票数据、拟合 Prophet 模型，并预测未来一段时间的股票价格。
}
\examples{
# 示例：对股票代码 "600048.SS" 的数据进行30天的预测
result <- stock_prediction_prophet(stock_data = stock_data_ss50[["600048.SS"]], prediction_period = 30)
print(result)

}
