% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{stock_treat}
\alias{stock_treat}
\title{股票交易资金模拟处理函数}
\usage{
stock_treat(
  buysell = c(3, 4, 5, -1, 5, 5, 0, 1, 2, 3, -1, 9, 7, 6, 7, 0, 2),
  increase = c(3.1, 0.4, -5, 0.9, 1.5, 5, 8, 1, 2, 3, 4, 1.9, 7, 6, 7, 6, 2),
  mymoney = 10000
)
}
\arguments{
\item{buysell}{一个数值向量，代表买卖信号。其中，-1 表示全仓买入信号，0 表示清仓信号，其他值可按需自定义但在本函数逻辑中不影响交易操作。}

\item{increase}{一个数值向量，与 `buysell` 长度相同，代表每个交易阶段对应的涨幅（百分比）。}

\item{mymoney}{初始资金，默认为 10000，是一个数值类型的值。}
}
\value{
一个数值向量，长度与 `buysell` 和 `increase` 相同，代表每个交易阶段结束后持有的资金总额。
}
\description{
此函数用于模拟股票交易过程中资金的变化情况。根据给定的买卖信号、涨幅数据和初始资金，
模拟资金在不同交易阶段的增长或减少，最终返回每个交易阶段结束后的资金总额。
}
\examples{
buysell <- c(3,4,5,-1,5,5,0,1,2,3,-1,9,7,6,7,0,2)
increase <- c(3.1,0.4,-5,0.9,1.5,5,8,1,2,3,4,1.9,7,6,7,6,2)
result <- stock_treat(buysell, increase, mymoney = 10000)
print(result)
}
