% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_stock.R
\name{volume_signal}
\alias{volume_signal}
\title{基于成交量突破的交易信号生成函数}
\usage{
volume_signal(stock_data, short_window = 10, long_window = 30)
}
\arguments{
\item{stock_data}{一个包含股票历史数据的 xts 对象，必须包含收盘价（Cl）和成交量（Vo）数据。}

\item{short_window}{短期成交量均线的窗口大小，默认值为5（即5日均线）。}

\item{long_window}{长期成交量均线的窗口大小，默认值为20（即20日均线）。}
}
\value{
返回一个数据框，其中包含日期、收盘价、成交量、短期和长期成交量均线，以及生成的交易信号。
        交易信号列的值为1表示买入信号，0表示卖出信号，NA表示没有信号。
}
\description{
该函数根据短期和长期成交量均线的突破生成买卖信号。若成交量突破均线，则生成买入信号；若成交量回落至均线以下，则生成卖出信号。
}
\examples{
# 假设你有一个包含历史数据的 stock_data 对象
# result <- volume_breakout_signal(stock_data)

}
